+++ scripts/helpers.js
@@ -2,8 +2,6 @@ import CONSTANTS from "./constants.js";
 import { getSetting } from "./lib/lib.js";
 
 export async function configureOneDndExhaustion() {
-  if (!getSetting(CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION)) return;
-  if (game.modules.get(CONSTANTS.MODULES.ALTERNATIVE_EXHAUSTION)?.active) return;
   if (getSetting(CONSTANTS.SETTINGS.ONE_DND_EXHAUSTION)) {
     if (CONFIG.DND5E.conditionTypes.exhaustion.levels !== 10) {
       if (!document.getElementById(CONSTANTS.STYLE_ELEMENT_ID)) {
@@ -32,29 +30,32 @@ export async function configureOneDndExhaustion() {
       CONFIG.DND5E.conditionEffects.halfMovement.delete("exhaustion-2");
       CONFIG.DND5E.conditionEffects.halfHealth.delete("exhaustion-4");
       CONFIG.DND5E.conditionEffects.noMovement.delete("exhaustion-5");
+      CONFIG.DND5E.conditionEffects.abilityCheckDisadvantage.delete("exhaustion-1");
+      CONFIG.DND5E.conditionEffects.abilitySaveDisadvantage.delete("exhaustion-3");
+      CONFIG.DND5E.conditionEffects.attackDisadvantage.delete("exhaustion-3");
       delete CONFIG.DND5E.conditionTypes.exhaustion.reduction;
     }
     if (game.modules.get("tidy5e-sheet")?.active) {
       await updateTidy5e();
     }
-  } else {
-    if (CONFIG.DND5E.conditionTypes.exhaustion.levels !== 6) {
-      let styleElement = document.getElementById(CONSTANTS.STYLE_ELEMENT_ID);
-      if (styleElement) {
-        styleElement.remove();
-      }
-      CONFIG.DND5E.conditionTypes.exhaustion.levels = 6;
-      CONFIG.DND5E.conditionTypes.exhaustion.icon = CONSTANTS.EXHAUSTION_CORE_PATH;
-      CONFIG.DND5E.conditionTypes.exhaustion.reference = CONFIG.DND5E.conditionTypes.exhaustion.oldReference ?? CONFIG.DND5E.conditionTypes.exhaustion.reference;
-      CONFIG.DND5E.conditionEffects.halfMovement.add("exhaustion-2");
-      CONFIG.DND5E.conditionEffects.halfHealth.add("exhaustion-4");
-      CONFIG.DND5E.conditionEffects.noMovement.add("exhaustion-5");
-      CONFIG.DND5E.conditionTypes.exhaustion.reduction = {rolls: 2, speed: 5};
-    }
-    if (game.modules.get("tidy5e-sheet")?.active) {
-      await updateTidy5e();
-    }
   }
+}
+
+export function getRestFlavor(durationIn, newDay, longRest) {
+  let duration;
+  let units;
+  if (durationIn % 1440 === 0) {
+    duration = durationIn / 1440;
+    units = duration > 1 ? 'Days' : 'Day';
+  } else if (durationIn % 60 === 0) {
+    duration = durationIn / 60;
+    units = duration > 1 ? 'Hours' : 'Hour';
+  } else {
+    duration = durationIn;
+    units = duration > 1 ? 'Minutes' : 'Minute';
+  }
+  return game.i18n.format(`REST-RECOVERY.Chat.Flavor.${longRest ? "Long" : "Short"}Rest${newDay ? "Newday" : "Normal"}`, {duration, units});
+  
 }
 
 export async function updateTidy5e() {
@@ -90,50 +91,11 @@ export async function updateTidy5e() {
 }
 
 export async function configureExhaustionHooks() {
-  let ac5eShouldControl = game.modules.get("automated-conditions-5e")?.active && game.settings.get("automated-conditions-5e", "autoExhaustion");
-  let alternativeExhaustionActive = game.modules.get(CONSTANTS.MODULES.ALTERNATIVE_EXHAUSTION)?.active;
-  let modernRules = game.settings.get("dnd5e", "rulesVersion") === "modern";
-  let disabledAutomation = !getSetting(CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION);
-  if (getSetting(CONSTANTS.SETTINGS.ONE_DND_EXHAUSTION) || ac5eShouldControl || alternativeExhaustionActive || modernRules || disabledAutomation) {
-    if (Hooks.events["dnd5e.preRollAbilityCheckV2"]) Hooks.off("dnd5e.preRollAbilityCheckV2", _preAbilityCheck);
-    if (Hooks.events["dnd5e.preRollSavingThrowV2"]) Hooks.off("dnd5e.preRollSavingThrowV2", _preAbilitySave);
-    if (Hooks.events["dnd5e.preRollDeathSaveV2"]) Hooks.off("dnd5e.preRollDeathSaveV2", _preDeathSave);
-    if (Hooks.events["dnd5e.preRollAttackV2"]) Hooks.off("dnd5e.preRollAttackV2", _preAttack);
-  } else {
-    // Take em off first just in case, that way we're always last
-    if (Hooks.events["dnd5e.preRollAbilityCheckV2"]) Hooks.off("dnd5e.preRollAbilityCheckV2", _preAbilityCheck);
-    if (Hooks.events["dnd5e.preRollSavingThrowV2"]) Hooks.off("dnd5e.preRollSavingThrowV2", _preAbilitySave);
-    if (Hooks.events["dnd5e.preRollDeathSaveV2"]) Hooks.off("dnd5e.preRollDeathSaveV2", _preDeathSave);
-    if (Hooks.events["dnd5e.preRollAttackV2"]) Hooks.off("dnd5e.preRollAttackV2", _preAttack);
-
-    Hooks.on("dnd5e.preRollAbilityCheckV2", _preAbilityCheck);
-    Hooks.on("dnd5e.preRollSavingThrowV2", _preAbilitySave);
-    Hooks.on("dnd5e.preRollDeathSaveV2", _preDeathSave);
-    Hooks.on("dnd5e.preRollAttackV2", _preAttack);
-  }
-}
-
-function _preAbilityCheck(config) {
-  if (!config.subject?.statuses.has("exhaustion")) return;
-  config.rolls[0].options.disadvantage = true;
-}
-
-function _preAbilitySave(config) {
-  if (!config.subject?.statuses?.has("exhaustion")) return;
-  if (!(config.subject?.system?.attributes?.exhaustion >= 3)) return;
-  config.rolls[0].options.disadvantage = true;
-}
-
-function _preDeathSave(config) {
-  if (!config.subject?.statuses?.has("exhaustion")) return;
-  if (!(config.subject?.system?.attributes?.exhaustion >= 3)) return;
-  config.rolls[0].options.disadvantage = true;
+  // Take em off first just in case, that way we're always last
+  if (Hooks.events["dnd5e.preRollAttackV2"]) Hooks.off("dnd5e.preRollAttackV2", _preAttack);
+  Hooks.on("dnd5e.preRollAttackV2", _preAttack);
 }
 
 function _preAttack(config) {
-  let actor = config.subject?.actor;
-  if (!actor) return;
-  if (!actor?.statuses?.has("exhaustion")) return;
-  if (!(actor?.system?.attributes?.exhaustion >= 3)) return;
-  config.rolls[0].options.disadvantage = true;
+  if (config.subject?.actor?.hasConditionEffect("attackDisadvantage")) config.rolls[0].options.disadvantage = true;
 }