+++ scripts/plugins.js
@@ -7,14 +7,10 @@ export default class plugins {
     const oneDndExhaustionEnabled = getSetting(CONSTANTS.SETTINGS.ONE_DND_EXHAUSTION);
     const exhaustionLevel = foundry.utils.getProperty(data, "system.attributes.exhaustion");
     const actorExhaustionEffect = actor.effects.find(effect => foundry.utils.getProperty(effect, "flags.rest-recovery.exhaustion-effect"));
-    if (!oneDndExhaustionEnabled && actorExhaustionEffect) {
-      return actor.deleteEmbeddedDocuments("ActiveEffect", [actorExhaustionEffect.id]);
-    }
-    if (!oneDndExhaustionEnabled) return;
+    if (!oneDndExhaustionEnabled) return actorExhaustionEffect?.delete();
     if (exhaustionLevel > 0 && !actorExhaustionEffect) {
-      oneDndExhaustionEffectData.label = game.i18n.localize("REST-RECOVERY.OneDnDExhaustionEffect.Name");
+      oneDndExhaustionEffectData.name = game.i18n.localize("REST-RECOVERY.OneDnDExhaustionEffect.Name");
       oneDndExhaustionEffectData.description = game.i18n.localize("REST-RECOVERY.OneDnDExhaustionEffect.Description");
-      oneDndExhaustionEffectData.flags.convenientDescription = game.i18n.localize("REST-RECOVERY.OneDnDExhaustionEffect.Description");
       return actor.createEmbeddedDocuments("ActiveEffect", [oneDndExhaustionEffectData]);
     } else if (exhaustionLevel <= 0 && actorExhaustionEffect) {
       return actor.deleteEmbeddedDocuments("ActiveEffect", [actorExhaustionEffect.id]);
@@ -33,8 +29,6 @@ const oneDndExhaustionEffectData = {
   "isDynamic": false,
   "isViewable": true,
   "flags": {
-    "isCustomConvenient": true,
-    "convenientDescription": "With One D&D exhaustion, a creature suffers cumulative -1 penalties per level of exhaustion to all d20 rolls (attack rolls, ability checks, skill checks, and saving throws) and spell DCs. When a creature reaches 10 levels of exhaustion, they die.",
     "rest-recovery": {
       "exhaustion-effect": true
     }